// lib/interface/trust-box.enums.ts
export enum ScriptInjectionStatus {
  Unknown = "unknown",
  Loading = "loading",
  Ready = "ready",
  Error = "error",
}

export enum TrustBoxWidgetType {
  Horizontal = "Horizontal",
  MicroButton = "MicroButton",
  MicroCombo = "MicroCombo",
  MicroReviewCount = "MicroReviewCount",
  MicroStar = "MicroStar",
  MicroTrustScore = "MicroTrustScore",
  Mini = "Mini",
  Starter = "Starter",
  ReviewCollector = "ReviewCollector",
  Carousel = "Carousel",
  DropDown = "DropDown",
  Grid = "Grid",
  List = "List",
  ListFiltered = "ListFiltered",
  MiniCarousel = "MiniCarousel",
  PopUp = "PopUp",
  Quote = "Quote",
  Slider = "Slider",
  ProductMini = "ProductMini",
  ProductMiniMultiSource = "ProductMiniMultiSource",
  ProductQA = "ProductQA",
  ProductReviews = "ProductReviews",
  ProductReviewsCarousel = "ProductReviewsCarousel",
  ProductReviewsGallery = "ProductReviewsGallery",
  ProductReviewsMultiSource = "ProductReviewsMultiSource",
  ProductReviewsMultiSourceSEO = "ProductReviewsMultiSourceSEO",
  ProductReviewsSEO = "ProductReviewsSEO",
}